User exclusion

+++ src/main/java/br/com/it3/model/dao/interfaces/BaseDAO.java
@@ -5,5 +5,5 @@ public interface BaseDAO<E> {
 
 	void remove(E entity);
 
-	E findById(Integer id);
+	E findById(long id);
 }
